Update reference to skin/WikiToLearnSkin and extensions/WikiToLearnVETemplates
This will trigger staging.
All the new environment templates will be uploaded.
The VisualEditor Template extension will be updated to support all the new templates.
The skin is now a little bit better on the SEO side and on Social Network.
The skin should now load a bit faster and content from CDN is preferred over locally hosted content.

Signed-off-by: Cristian Baldi <bld.cris.96@gmail.com>